+++ spinta/commands/read.py
@@ -1,4 +1,5 @@
 from starlette.requests import Request
+from starlette.responses import Response
 
 from spinta import commands
 from spinta.backends.helpers import get_select_prop_names
@@ -104,13 +105,34 @@ async def getone(
     *,
     action: Action,
     params: UrlParams,
-):
+) -> Response:
     if params.prop and params.propref:
-        return await commands.getone(context, request, params.prop, params.model.backend, action=action, params=params)
+        return await commands.getone(
+            context,
+            request,
+            params.prop,
+            params.model.backend,
+            action=action,
+            params=params,
+        )
     elif params.prop:
-        return await commands.getone(context, request, params.prop, params.prop.dtype.backend or params.model.backend, action=action, params=params)
+        return await commands.getone(
+            context,
+            request,
+            params.prop,
+            params.prop.dtype.backend or params.model.backend,
+            action=action,
+            params=params,
+        )
     else:
-        return await commands.getone(context, request, params.model, params.model.backend, action=action, params=params)
+        return await commands.getone(
+            context,
+            request,
+            params.model,
+            params.model.backend,
+            action=action,
+            params=params,
+        )
 
 
 @commands.changes.register(Context, Model, Request)