+++ tests/test_urlparams.py
@@ -1,30 +1,24 @@
-from starlette.requests import Request
-
-from spinta.components import Context, UrlParams, Version
+from spinta.components import Context
+from spinta.components import UrlParams
+from spinta.components import Version
 from spinta.commands import prepare
+from spinta.testing.request import make_get_request
 
 
-def _parse(context: Context, url: str, accept='application/json') -> UrlParams:
-    params = UrlParams()
-    if '?' in url:
-        path, query = url.split('?')
-    else:
-        path, query = url, ''
-    request = Request({
-        'type': 'http',
-        'path': path,
-        'path_params': {
-            'path': path,
-        },
-        'query_string': query.encode(),
-        'headers': [
-            (b'Accept', accept.encode()),
-        ]
-    })
-    return prepare(context, params, Version(), request)
+def _parse(
+    context: Context,
+    query: str,
+    accept='application/json',
+) -> UrlParams:
+    request = make_get_request('', query, {'Accept': accept})
+    return prepare(context, UrlParams(), Version(), request)
 
 
 def test_format(context):
-    assert _parse(context, '?format(width(42))').formatparams == {'width': 42}
-    assert _parse(context, '?format(csv,width(42))').format == 'csv'
-    assert _parse(context, '?format(csv,width(42))').formatparams == {'width': 42}
+    assert _parse(context, 'format(csv,width(42))').format == 'csv'
+    assert _parse(context, 'format(width(42))').formatparams == {
+        'width': 42,
+    }
+    assert _parse(context, 'format(csv,width(42))').formatparams == {
+        'width': 42,
+    }